Updates 3.0 and cleaning before handing-in
