+++ src/main/java/org/matsim/prepare/superblocks/ScenarioB25/NetworkModifierScenarioB25.java
@@ -1,4 +1,4 @@
-package org.matsim.prepare.superblocks.ScenarioA75;
+package org.matsim.prepare.superblocks.ScenarioB25;
 
 import org.apache.log4j.Logger;
 import org.locationtech.jts.geom.Geometry;
@@ -20,21 +20,21 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.*;
 
-public class NetworkModifierScenarioC75 {
+public class NetworkModifierScenarioB25 {
 
-    private static final Logger LOG = Logger.getLogger(org.matsim.prepare.superblocks.ScenarioA75.NetworkModifierScenarioC75.class);
+    private static final Logger LOG = Logger.getLogger(NetworkModifierScenarioB25.class);
 
     public static void main(String[] args) throws IOException {
         // Input and output files
         String networkInputFile = "/Users/moritzkreuschner/Desktop/Master Thesis/01_Shapefiles/Shapefiles/berlin-v5.5-network.xml.gz";
-        String networkOutputFile = "/Users/moritzkreuschner/Desktop/Master Thesis/02_Coding/git/matsim-berlin-kreuschner/superblock_input_data/Network-modifiedC75.xml.gz";
+        String networkOutputFile = "/Users/moritzkreuschner/Desktop/Master Thesis/02_Coding/git/matsim-berlin-kreuschner/superblock_input_data/Network-modifiedA25.xml.gz";
 
-        Path filePath = Paths.get("/Users/moritzkreuschner/Desktop/Master Thesis/01_Shapefiles/Shapefiles/Superblocks_Shapefiles/75percent/NOTin75percent.txt");
+        Path filePath = Paths.get("/Users/moritzkreuschner/Desktop/Master Thesis/01_Shapefiles/Shapefiles/Superblocks_Shapefiles/25percent/NOTin25percent.txt");
         Scanner scanner = new Scanner(filePath);
-        List<Integer> NOTin75list = new ArrayList<>();
+        List<Integer> NOTin25list = new ArrayList<>();
         while (scanner.hasNext()) {
             if (scanner.hasNextInt()) {
-                NOTin75list.add(scanner.nextInt());
+                NOTin25list.add(scanner.nextInt());
             } else {
                 scanner.next();
             }
@@ -45,7 +45,6 @@ public class NetworkModifierScenarioC75 {
         MatsimNetworkReader reader = new MatsimNetworkReader(scenario.getNetwork());
         reader.readFile(networkInputFile);
 
-        // Get only non-pt links
         Set<Link> set = new HashSet<>();
         for (Link link1 : scenario.getNetwork().getLinks().values()) {
             if (!link1.getAllowedModes().contains(TransportMode.pt)) {
@@ -55,10 +54,9 @@ public class NetworkModifierScenarioC75 {
         Set<? extends Link> nonptlinks = set;
 
         // Loop for different shapefiles
-        for (int i = 1; i < 160; i++) {
-
-            // Superblocks that are not in the directory
-            if (NOTin75list.contains(i)) {
+        // Superblocks that are not in the directory
+        for (int i = 1; i < 161; i++) {
+            if (NOTin25list.contains(i)) {
                 continue;
             } else {
 
@@ -77,25 +75,30 @@ public class NetworkModifierScenarioC75 {
                 for (Link link : nonptlinks) {
 
                     Point linkCenterAsPoint = MGC.xy2Point(link.getCoord().getX(), link.getCoord().getY());
-
                     if (areaGeometry.contains(linkCenterAsPoint)) {
-                        link.setFreespeed(1.3888889);
+                        link.setFreespeed(0.00001);
+                        link.setCapacity(0);
                     }
                 }
 
                 LOG.info("Superblock " + i + " is ready");
             }
         }
-        LOG.info("Finished modifying freespeed");
+
+        LOG.info("Finished modifying freespeed and capacity");
 
         // Get car subnetwork and clean it
         Scenario carScenario = ScenarioUtils.createScenario(ConfigUtils.createConfig());
         new MultimodalNetworkCleaner(carScenario.getNetwork()).run(Set.of(TransportMode.car));
+        /*TransportModeNetworkFilter transportModeNetworkFilterCar = new TransportModeNetworkFilter(scenario.getNetwork());
+        transportModeNetworkFilterCar.filter(carScenario.getNetwork(), new HashSet<>(Arrays.asList(TransportMode.car)));
+        (new NetworkCleaner()).run(carScenario.getNetwork());*/
 
-        LOG.info("Finished creating and cleaning car subnetwork");
+        LOG.info("Finished cleaning car subnetwork");
 
         // Write modified network to file
         NetworkWriter writer = new NetworkWriter(scenario.getNetwork());
         writer.write(networkOutputFile);
+
     }
 }
